requirements.txt
# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
scrapy==2.11.0
praw==7.7.1  # Reddit API
tweepy==4.14.0  # Twitter API

# Data Processing
pandas==2.2.0
numpy==1.26.3
nltk==3.8.1

# ML & LLM
torch==2.2.0
transformers==4.37.2
peft==0.8.2  # LoRA fine-tuning
accelerate==0.26.1
bitsandbytes==0.42.0  # Quantization
datasets==2.16.1

# Embeddings & Vector DB
sentence-transformers==2.3.1
chromadb==0.4.22

# Evaluation
evaluate==0.4.1
nltk==3.8.1
rouge-score==0.1.2

# Database
sqlite3  # Built-in

# API & Backend
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0

# Task Scheduling
apscheduler==3.10.4

# Demo Interface
streamlit==1.31.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1
wandb==0.16.3  # Experiment tracking

# Development
jupyter==1.0.0
ipywidgets==8.1.1
matplotlib==3.8.2
seaborn==0.13.2
